move utils to /shared
